Fix minimap integration after rebase
